[APPC-1247] Extracted duplicated textwatcher from sms validation
